add quest adjust strState part

+++ ci/quest/utility/quest.py
@@ -27,16 +27,14 @@ class QuestUtility:
     def deleteQuest(self, ciuserRequest=None, strQID=None):
         questTarget = self.getCIQuestByQID(strQID=strQID)
         #確認是否為 發起人
-        if questTarget.ciuserInitiator == ciuserRequest:
+        if questTarget.ciuserInitiator == ciuserRequest and questTarget.strState == "new":
             questTarget.delete()
     
     #接受任務
     def acceptQuest(self, ciuserRequest=None, strQID=None):
         questTarget = self.getCIQuestByQID(strQID=strQID)
         #確認是否為 除發起人以外的人
-        if questTarget.ciuserInitiator == ciuserRequest:
-            pass
-        else:
+        if questTarget.ciuserInitiator != ciuserRequest and questTarget.strState == "new":
             #儲存 ciuserExecutor 及 strState
             questTarget.ciuserExecutor = ciuserRequest
             questTarget.strState = "processing"
@@ -46,7 +44,7 @@ class QuestUtility:
     def abandonQuest(self, ciuserRequest=None, strQID=None):
         questTarget = self.getCIQuestByQID(strQID=strQID)
         #確認是否為 執行人
-        if questTarget.ciuserExecutor == ciuserRequest:
+        if questTarget.ciuserExecutor == ciuserRequest and questTarget.strState == "processing":
             #儲存 ciuserExecutor 及 strState
             questTarget.ciuserExecutor = None
             questTarget.strState = "new"
